+++ src/mesh/meshblock.hpp
@@ -1,0 +1,314 @@
+//========================================================================================
+// Athena++ astrophysical MHD code
+// Copyright(C) 2014 James M. Stone <jmstone@princeton.edu> and other code contributors
+// Licensed under the 3-clause BSD License, see LICENSE file for details
+//========================================================================================
+// (C) (or copyright) 2020. Triad National Security, LLC. All rights reserved.
+//
+// This program was produced under U.S. Government contract 89233218CNA000001 for Los
+// Alamos National Laboratory (LANL), which is operated by Triad National Security, LLC
+// for the U.S. Department of Energy/National Nuclear Security Administration. All rights
+// in the program are reserved by Triad National Security, LLC, and the U.S. Department
+// of Energy/National Nuclear Security Administration. The Government is granted for
+// itself and others acting on its behalf a nonexclusive, paid-up, irrevocable worldwide
+// license in this material to reproduce, prepare derivative works, distribute copies to
+// the public, perform publicly and display publicly, and to permit others to do so.
+//========================================================================================
+#ifndef MESH_MESHBLOCK_HPP_
+#define MESH_MESHBLOCK_HPP_
+
+#include <cstdint>
+#include <functional>
+#include <map>
+#include <memory>
+#include <string>
+#include <vector>
+
+#include "application_input.hpp"
+#include "bvals/bvals.hpp"
+#include "bvals/bvals_interfaces.hpp"
+#include "config.hpp"
+#include "coordinates/coordinates.hpp"
+#include "defs.hpp"
+#include "domain.hpp"
+#include "interface/container.hpp"
+#include "interface/container_collection.hpp"
+#include "interface/properties_interface.hpp"
+#include "interface/state_descriptor.hpp"
+#include "interface/update.hpp"
+#include "kokkos_abstraction.hpp"
+#include "mesh/mesh_refinement.hpp"
+#include "mesh/meshblock.hpp"
+#include "mesh/meshblock_tree.hpp"
+#include "outputs/io_wrapper.hpp"
+#include "parameter_input.hpp"
+#include "parthenon_arrays.hpp"
+#include "reconstruct/reconstruction.hpp"
+#include "utils/interp_table.hpp"
+
+namespace parthenon {
+
+// Forward declarations
+class BoundaryValues;
+class Mesh;
+class MeshBlockTree;
+class MeshRefinement;
+class ParameterInput;
+class Reconstruction;
+
+// Inner loop default pattern
+// - Defined outside of the MeshBlock class because it does not require an exec space
+// - Not defined in kokkos_abstraction.hpp because it requires the compile time option
+//   DEFAULT_INNER_LOOP_PATTERN to be set.
+template <typename Function>
+KOKKOS_INLINE_FUNCTION void par_for_inner(const team_mbr_t &team_member, const int &il,
+                                          const int &iu, const Function &function) {
+  parthenon::par_for_inner(DEFAULT_INNER_LOOP_PATTERN, team_member, il, iu, function);
+}
+
+//----------------------------------------------------------------------------------------
+//! \class MeshBlock
+//  \brief data/functions associated with a single block
+class MeshBlock {
+  friend class RestartOutput;
+  friend class Mesh;
+
+ public:
+  MeshBlock(const int n_side, const int ndim); // for Kokkos testing with ghost
+  MeshBlock(int igid, int ilid, LogicalLocation iloc, RegionSize input_size,
+            BoundaryFlag *input_bcs, Mesh *pm, ParameterInput *pin,
+            ApplicationInput *app_in, Properties_t &properties, int igflag,
+            bool ref_flag = false);
+  MeshBlock(int igid, int ilid, LogicalLocation iloc, RegionSize input_block,
+            BoundaryFlag *input_bcs, Mesh *pm, ParameterInput *pin,
+            ApplicationInput *app_in, Properties_t &properties, Packages_t &packages,
+            int igflag, bool ref_flag = false);
+  MeshBlock(int igid, int ilid, Mesh *pm, ParameterInput *pin, ApplicationInput *app_in,
+            Properties_t &properties, Packages_t &packages, LogicalLocation iloc,
+            RegionSize input_block, BoundaryFlag *input_bcs, double icost, int igflag);
+  ~MeshBlock();
+
+  // Kokkos execution space for this MeshBlock
+  DevExecSpace exec_space;
+
+  // data
+  Mesh *pmy_mesh = nullptr; // ptr to Mesh containing this MeshBlock
+  LogicalLocation loc;
+  RegionSize block_size;
+  // for convenience: "max" # of real+ghost cells along each dir for allocating "standard"
+  // sized MeshBlock arrays, depending on ndim i.e.
+  //
+  // cellbounds.nx2 =    nx2      + 2*NGHOST if   nx2 > 1
+  // (entire)         (interior)               (interior)
+  //
+  // Assuming we have a block cells, and nx2 = 6, and NGHOST = 1
+  //
+  // <----- nx1 = 8 ---->
+  //       (entire)
+  //
+  //     <- nx1 = 6 ->
+  //       (interior)
+  //
+  //  - - - - - - - - - -   ^
+  //  |  |  ghost    |  |   |
+  //  - - - - - - - - - -   |         ^
+  //  |  |     ^     |  |   |         |
+  //  |  |     |     |  |  nx2 = 8    nx2 = 6
+  //  |  | interior  |  | (entire)   (interior)
+  //  |  |     |     |  |             |
+  //  |  |     v     |  |   |         v
+  //  - - - - - - - - - -   |
+  //  |  |           |  |   |
+  //  - - - - - - - - - -   v
+  //
+  IndexShape cellbounds;
+  // on 1x coarser level MeshBlock i.e.
+  //
+  // c_cellbounds.nx2 = cellbounds.nx2 * 1/2 + 2*NGHOST, if  cellbounds.nx2 >1
+  //   (entire)             (interior)                          (interior)
+  //
+  // Assuming we have a block cells, and nx2 = 6, and NGHOST = 1
+  //
+  //          cells                              c_cells
+  //
+  //  - - - - - - - - - -   ^              - - - - - - - - - -     ^
+  //  |  |           |  |   |              |  |           |  |     |
+  //  - - - - - - - - - -   |              - - - - - - - - - -     |
+  //  |  |     ^     |  |   |              |  |      ^    |  |     |
+  //  |  |     |     |  |   |              |  |      |    |  |     |
+  //  |  |  nx2 = 6  |  |  nx2 = 8  ====>  |  |   nx2 = 3 |  |   nx2 = 5
+  //  |  |(interior) |  |  (entire)        |  | (interior)|  |  (entire)
+  //  |  |     v     |  |   |              |  |      v    |  |     |
+  //  - - - - - - - - - -   |              - - - - - - - - - -     |
+  //  |  |           |  |   |              |  |           |  |     |
+  //  - - - - - - - - - -   v              - - - - - - - - - -     v
+  //
+  IndexShape c_cellbounds;
+  int gid, lid;
+  int cnghost;
+  int gflag;
+
+  // The User defined containers
+  ContainerCollection<Real> real_containers;
+
+  Properties_t properties;
+  Packages_t packages;
+
+  std::unique_ptr<MeshBlockApplicationData> app;
+
+  Coordinates_t coords;
+
+  // mesh-related objects
+  // TODO(jcd): remove all these?
+  std::unique_ptr<BoundaryValues> pbval;
+  std::unique_ptr<MeshRefinement> pmr;
+  std::unique_ptr<Reconstruction> precon;
+
+  BoundaryFlag boundary_flag[6];
+
+  // functions
+
+  //----------------------------------------------------------------------------------------
+  //! \fn void MeshBlock::DeepCopy(const DstType& dst, const SrcType& src)
+  //  \brief Deep copy between views using the exec space of the MeshBlock
+  template <class DstType, class SrcType>
+  void deep_copy(const DstType &dst, const SrcType &src) {
+    Kokkos::deep_copy(exec_space, dst, src);
+  }
+
+  // 1D default loop pattern
+  template <typename Function>
+  inline void par_for(const std::string &name, const int &il, const int &iu,
+                      const Function &function) {
+    // using loop_pattern_flatrange_tag instead of DEFAULT_LOOP_PATTERN for now
+    // as the other wrappers are not implemented yet for 1D loops
+    parthenon::par_for(loop_pattern_flatrange_tag, name, exec_space, il, iu, function);
+  }
+
+  // 2D default loop pattern
+  template <typename Function>
+  inline void par_for(const std::string &name, const int &jl, const int &ju,
+                      const int &il, const int &iu, const Function &function) {
+    // using loop_pattern_mdrange_tag instead of DEFAULT_LOOP_PATTERN for now
+    // as the other wrappers are not implemented yet for 1D loops
+    parthenon::par_for(loop_pattern_mdrange_tag, name, exec_space, jl, ju, il, iu,
+                       function);
+  }
+
+  // 3D default loop pattern
+  template <typename Function>
+  inline void par_for(const std::string &name, const int &kl, const int &ku,
+                      const int &jl, const int &ju, const int &il, const int &iu,
+                      const Function &function) {
+    parthenon::par_for(DEFAULT_LOOP_PATTERN, name, exec_space, kl, ku, jl, ju, il, iu,
+                       function);
+  }
+
+  // 4D default loop pattern
+  template <typename Function>
+  inline void par_for(const std::string &name, const int &nl, const int &nu,
+                      const int &kl, const int &ku, const int &jl, const int &ju,
+                      const int &il, const int &iu, const Function &function) {
+    parthenon::par_for(DEFAULT_LOOP_PATTERN, name, exec_space, nl, nu, kl, ku, jl, ju, il,
+                       iu, function);
+  }
+
+  // 1D Outer default loop pattern
+  template <typename Function>
+  inline void par_for_outer(const std::string &name, const size_t &scratch_size_in_bytes,
+                            const int &scratch_level, const int &kl, const int &ku,
+                            const Function &function) {
+    parthenon::par_for_outer(DEFAULT_OUTER_LOOP_PATTERN, name, exec_space,
+                             scratch_size_in_bytes, scratch_level, kl, ku, function);
+  }
+  // 2D Outer default loop pattern
+  template <typename Function>
+  inline void par_for_outer(const std::string &name, const size_t &scratch_size_in_bytes,
+                            const int &scratch_level, const int &kl, const int &ku,
+                            const int &jl, const int &ju, const Function &function) {
+    parthenon::par_for_outer(DEFAULT_OUTER_LOOP_PATTERN, name, exec_space,
+                             scratch_size_in_bytes, scratch_level, kl, ku, jl, ju,
+                             function);
+  }
+
+  // 3D Outer default loop pattern
+  template <typename Function>
+  inline void par_for(const std::string &name, size_t &scratch_size_in_bytes,
+                      const int &scratch_level, const int &nl, const int &nu,
+                      const int &kl, const int &ku, const int &jl, const int &ju,
+                      const Function &function) {
+    parthenon::par_for_outer(DEFAULT_OUTER_LOOP_PATTERN, name, exec_space,
+                             scratch_size_in_bytes, scratch_level, nl, nu, kl, ku, jl, ju,
+                             function);
+  }
+
+  // Inner loop default pattern
+  template <typename Function>
+  KOKKOS_INLINE_FUNCTION void par_for_inner(const team_mbr_t &team_member, const int &il,
+                                            const int &iu, const Function &function) {
+    parthenon::par_for_inner(DEFAULT_INNER_LOOP_PATTERN, team_member, il, iu, function);
+  }
+
+  std::size_t GetBlockSizeInBytes();
+  int GetNumberOfMeshBlockCells() const {
+    return block_size.nx1 * block_size.nx2 * block_size.nx3;
+  }
+  void SearchAndSetNeighbors(MeshBlockTree &tree, int *ranklist, int *nslist);
+  void WeightedAve(ParArrayND<Real> &u_out, ParArrayND<Real> &u_in1,
+                   ParArrayND<Real> &u_in2, const Real wght[3]);
+  void WeightedAve(FaceField &b_out, FaceField &b_in1, FaceField &b_in2,
+                   const Real wght[3]);
+
+  void ResetToIC() { ProblemGenerator(nullptr, nullptr); }
+
+  // inform MeshBlock which arrays contained in member Field, Particles,
+  // ... etc. classes are the "primary" representations of a quantity. when registered,
+  // that data are used for (1) load balancing (2) (future) dumping to restart file
+  void RegisterMeshBlockData(std::shared_ptr<CellVariable<Real>> pvar_cc);
+  void RegisterMeshBlockData(std::shared_ptr<FaceField> pvar_fc);
+
+  // defined in either the prob file or default_pgen.cpp in ../pgen/
+  static void
+  UserWorkBeforeOutputDefault(ParameterInput *pin); // called in Mesh fn (friend class)
+  std::function<void(ParameterInput *)> UserWorkBeforeOutput =
+      &UserWorkBeforeOutputDefault;
+  static void UserWorkInLoopDefault(); // called in TimeIntegratorTaskList
+  std::function<void()> UserWorkInLoop = &UserWorkInLoopDefault;
+  void SetBlockTimestep(const Real dt) { new_block_dt_ = dt; }
+  Real NewDt() const { return new_block_dt_; }
+
+ private:
+  // data
+  Real new_block_dt_, new_block_dt_hyperbolic_, new_block_dt_parabolic_,
+      new_block_dt_user_;
+  std::vector<std::shared_ptr<CellVariable<Real>>> vars_cc_;
+  std::vector<std::shared_ptr<FaceField>> vars_fc_;
+
+  void InitializeIndexShapes(const int nx1, const int nx2, const int nx3);
+  // functions
+  void SetCostForLoadBalancing(double cost);
+
+  // defined in either the prob file or default_pgen.cpp in ../pgen/
+  static void ProblemGeneratorDefault(MeshBlock *pmb, ParameterInput *pin);
+  std::function<void(MeshBlock *, ParameterInput *)> ProblemGenerator =
+      &ProblemGeneratorDefault;
+  static pMeshBlockApplicationData_t
+  InitApplicationMeshBlockDataDefault(ParameterInput *pin);
+  std::function<pMeshBlockApplicationData_t(ParameterInput *)>
+      InitApplicationMeshBlockData = &InitApplicationMeshBlockDataDefault;
+  static void InitUserMeshBlockDataDefault(ParameterInput *pin);
+  std::function<void(ParameterInput *)> InitUserMeshBlockData =
+      &InitUserMeshBlockDataDefault;
+
+  // functions and variables for automatic load balancing based on timing
+  Kokkos::Timer lb_timer;
+  double cost_;
+  void ResetTimeMeasurement();
+  void StartTimeMeasurement();
+  void StopTimeMeasurement();
+};
+using BlockList_t = std::vector<std::shared_ptr<MeshBlock>>;
+
+} // namespace parthenon
+
+#endif // MESH_MESHBLOCK_HPP_
